address to link

+++ Memento/Views/ListView.swift
@@ -22,38 +22,22 @@ struct ListView: View {
             List(filteredItems) { item in
                 ItemView(item: item)
                     .contextMenu(ContextMenu(menuItems: {
-                        Button("Delete", systemImage: "trash", role: .destructive) {
-                            deleteItem(item: item)
-                            UpdateAll()
-                        }
+                        Button("Delete", systemImage: "trash", role: .destructive, action: {deleteItem(item: item)})
                         Button(item.viewed ? "Unmark Viewed": "Mark Viewed", systemImage: "book") {
                             item.viewed.toggle()
                             UpdateAll()
                         }
                     }))
                     .swipeActions(edge: .leading) {
-                        Button {
-                            item.viewed.toggle()
-                            UpdateAll()
-                        } label: {
-                            Label({item.viewed ? "Unmark Viewed" : "Mark Viewed"}(), systemImage: "book")
-                        }
-                        .tint(.indigo)
+                        Button({item.viewed ? "Unmark Viewed" : "Mark Viewed"}(), systemImage: "book", action: {viewModel.toggleViewed(item)}).tint(.indigo)
                     }
                     .swipeActions(edge: .trailing) {
-                        Button(role: .destructive) {
-                            withAnimation {
-                                modelContext.delete(item)
-                                UpdateAll()
-                            }
-                        } label: {
-                            Label("Delete", systemImage: "trash")
-                        }
+                        Button("Delete", systemImage: "trash", role: .destructive, action: {viewModel.toggleViewed(item)})
                     }
             }
             .searchable(text: $viewModel.searchText, prompt: "Search Items")
             .overlay {
-                if filteredItems.isEmpty {
+                if viewModel.filterItems(items).isEmpty {
                     if viewModel.searchText.isEmpty {
                         ContentUnavailableView("No Items Added", systemImage: "doc")
                     } else {
@@ -61,20 +45,7 @@ struct ListView: View {
                     }
                 }
             }
-            .toolbar {
-                ToolbarItem {
-                    Button("Add Item From Clipboard", systemImage: "clipboard") {
-                        Task {
-                            await addFromPaste()
-                        }
-                    }
-                }
-                ToolbarItem {
-                    Button(action: viewModel.addSheet) {
-                        Label("Add Item", systemImage: "plus")
-                    }
-                }
-            }
+            .toolbar { Button("Add Item", systemImage: "plus", action: viewModel.addSheet) }
             .navigationTitle("Items")
             .sheet(isPresented: $viewModel.sheetShown, content: {
                 AddView(shown: $viewModel.sheetShown)
@@ -91,10 +62,10 @@ struct ListView: View {
     
     func addFromPaste() async {
         let pasteText = paste()
-        guard let address: String = pasteText else {
+        guard let link: String = pasteText else {
             return
         }
-        guard let item = await makeItem(address: address) else {
+        guard let item = await makeItem(link: link) else {
             return
         }
         withAnimation {
